backend_portable/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.6.0

# LangChain + LangGraph
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0
langgraph>=0.2.0

# Vector store & embeddings
chromadb>=0.5.0
sentence-transformers>=3.3.0

# HTTP client
httpx>=0.27.0

# PDF parsing
pypdf>=5.1.0

# Utilities
python-multipart>=0.0.12
uuid>=1.30
